fix: centralized both User and BookModel in db folder
