+++ src/db/models.py
@@ -1,9 +1,36 @@
 import uuid
-from typing import Optional
-from src.auth import models
+from typing import List, Optional
 from datetime import datetime, date
 import sqlalchemy.dialects.postgresql as pg
 from sqlmodel import SQLModel, Field, Column, Relationship
+
+
+class User(SQLModel, table=True):
+    __tablename__: str = "user"
+    
+    uid: uuid.UUID = Field(
+        sa_column = Column(
+            pg.UUID, 
+            nullable=False, 
+            primary_key=True, 
+            default=uuid.uuid4
+            )
+        )
+    created_at: datetime = Field(sa_column=Column(pg.TIMESTAMP, default=datetime.now))
+    updated_at: datetime = Field(sa_column=Column(pg.TIMESTAMP, default=datetime.now))
+    username: str
+    email: str
+    role: str = Field(sa_column=Column(pg.VARCHAR, nullable=False, server_default="user"))
+    password: str # = Field(exclude=True) # won't show password in response
+    first_name: str
+    last_name: str
+    is_verified: bool = Field(default=False)
+    books: List["BookModel"] = Relationship(back_populates="user",sa_relationship_kwargs={"lazy":"selectin"})
+
+
+    """ String representation of the User object """
+    def __repr__(self):
+        return f"<User {self.username}>"
 
 
 # Create Book
@@ -27,7 +54,7 @@ class BookModel(SQLModel, table=True):
     language: Optional[str] = "English"
     published_date: Optional[date] = None
     user_uid: Optional[uuid.UUID] = Field(default=None, foreign_key="user.uid")
-    user: Optional["models.User"] = Relationship(back_populates="books")
+    user: Optional["User"] = Relationship(back_populates="books")
 
     def __repr__(self):
         return f"BookModel({self.title}, {self.author}, {self.publisher}, {self.page_count}, {self.language}, {self.published_date})"
